test: replace `ts-node` with `esbuild` as jest config loader
Resolves an issue where Jest fails to parse its TypeScript config file, giving the following error:

> Error: Jest: Failed to parse the TypeScript config file ./jest.config.ts
>   TypeError: value.replace is not a function

For an undetermined reason, this issue only occurs in CI (or at least inside a Docker image
simulating the CI environment). It was not occuring when running tests on my local machine.

A [comment][1] on [jest#14096][2] points to an issue with `ts-node`, the loader used to load the
config file: https://github.com/TypeStrong/ts-node/issues/1954. This issue requests support for
array values for the `extends` config in `tsconfig.json`. Testing the `extends` config being a
non-array value does stop the above error, verifying it is an issue with `ts-node`.

An attempt to use the `ts-node` `11.0.0` beta did not resolve the issue as a different error was
thrown:

> Error: Jest: Failed to parse the TypeScript config file ./jest.config.ts
>   TypeError: state.conditions.includes is not a function

This was not investigated further. Instead, switching the loader to `esbuild-register`, which is
the other loader supported by Jest (see [documentation][3]), resolves the issue and avoids sinking
time into investigating.

[1]: https://github.com/jestjs/jest/issues/14096#issuecomment-1518583709
[2]: https://github.com/jestjs/jest/issues/14096
[3]: https://jestjs.io/docs/configuration